Add test to ExpressionParserTests reproducing problems with where function
Functions which perform aggregations are a problem when run within the context of the argument to a where function, as they remove columns, including the input value which we rely upon to produce the where result.

We need to look at the three functions that do aggregation, and ensure that the input value column is preserved when there is a "this context".

1. AbstractAggregateFunction
2. MembershipOperator
3. SubsumesFunction

Reproduces #34.