Use dependency injection for test dependencies
This ensures that we only have one way of managing things like
SparkSession and FhirContext across the tests, and the application.

We now use Spring profiles to distinguish between the different run
profiles, and to filter beans accordingly.

We have also changed the way we bring in configuration, to be profile
aware.

Tests are now split into three distinct levels:

1. Unit tests (which are isolated to particular parts of the app)
2. Integration tests (which run the whole server, and interrogate the
   FHIR API)
3. System tests (which run up the Docker image and fire requests at it)

Resolves #232.